chore(infrastructure): move workspaces collections to mongodb
