+++ src/kairos/shared/infrastructure/mongodb/collection/workspaces-collection.ts
@@ -1,0 +1,7 @@
+import { mongoDBClient } from '@/kairos/shared/persistence/mongodb/mongodb-composition';
+import {
+  WorkspacesCollection,
+  workspacesCollectionName,
+} from '@/kairos/shared/persistence/mongodb/workspaces-collection-schema';
+
+export const workspacesCollection: WorkspacesCollection = mongoDBClient.db().collection(workspacesCollectionName);
